requirements.txt
pytubefix
rich
attrs
pydash
future
validators
flet